+++ Package.swift
@@ -1,30 +1,37 @@
-// swift-tools-version:5.3
+// swift-tools-version:6.0
 import PackageDescription
 
 let package = Package(
     name: "dogstatsd",
     platforms: [
-       .macOS(.v10_15)
+        .macOS(.v10_15)
     ],
     products: [
-            // Products define the executables and libraries produced by a package, and make them visible to other packages.
-            .library(
-                name: "Dogstatsd",
-                targets: ["Dogstatsd"]),
-        ],
+        .library(
+            name: "Dogstatsd",
+            targets: ["Dogstatsd"]
+        ),
+    ],
     dependencies: [
-        .package(url: "https://github.com/vapor/vapor.git", from: "4.0.0"),
+        .package(url: "https://github.com/vapor/vapor.git", exact: "4.121.4"),
     ],
     targets: [
         .target(
             name: "Dogstatsd",
             dependencies: [
-                .product(name: "Vapor", package: "vapor")
+                .product(name: "Vapor", package: "vapor"),
             ]
         ),
-        .testTarget(name: "Tests", dependencies: [
-            .target(name: "Dogstatsd"),
-            .product(name: "XCTVapor", package: "vapor"),
-        ])
+        .testTarget(
+            name: "DogstatsdTests",
+            dependencies: [
+                .target(name: "Dogstatsd"),
+                .product(name: "XCTVapor", package: "vapor"),
+            ],
+            path: "Tests/DogstatsdTests"
+        )
+    ],
+    swiftLanguageModes: [
+        .v6
     ]
 )